refactor(invoices): extract authenticated PDF artifact lifecycle
