+++ backend/app/features/invoices/pdf_artifacts/__init__.py
@@ -1,0 +1,5 @@
+"""Invoice authenticated PDF artifact lifecycle slice."""
+
+from app.features.invoices.pdf_artifacts.service import InvoicePdfArtifactService
+
+__all__ = ["InvoicePdfArtifactService"]
